pyproject.toml
[tool.poetry]
name = "ai-learning-backend"
version = "1.0.0"
description = ""
authors = ["Gaurav Aggarwal"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"
uvicorn = "^0.20.0"
pydantic = "^2.9.2"
google-generativeai = "^0.8.1"
pydantic-settings = "^2.5.2"
redis = "^5.0.8"
minio = "^7.2.8"
bcrypt = "^4.2.0"
motor = "^3.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.10"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
